Fix libedac.so library version in 0.16 release
The libtool AGE of libedac.so was incremented for the 0.16
release of edac-utils without a corresponding increase in
CURRENT, which caused the libedac.so libtool version to
decrement from 1.0.0 to 0.1.0.

Instead, increment both AGE and CURRENT, since an interface
was changed, but with backwards compatibility, and
prepare to tag a 0.16.1.